移除lombok，否则java17中会报错

+++ common/src/test/java/io/seata/common/util/BeanUtilsTest.java
@@ -74,7 +74,7 @@ public class BeanUtilsTest {
         Assertions.assertEquals(branchDO.getTransactionId(), Long.valueOf(map.get("transactionId")));
         Assertions.assertEquals(branchDO.getStatus(), Integer.valueOf(map.get("status")));
         Assertions.assertEquals(branchDO.getTest(), Double.valueOf(map.get("test")));
-        Assertions.assertEquals(branchDO.getGmtCreate().getTime(), Long.valueOf(map.get("gmtCreate")).longValue());
+        Assertions.assertEquals(branchDO.getGmtCreate().getTime(),Long.valueOf(map.get("gmtCreate")));
     }
 
     @Test
